requirements.txt
chainlit==1.3.2
anthropic==0.30.0
langgraph==0.2.14
langchain-anthropic==0.1.23
langchain-core==0.2.38
openpyxl==3.1.5
pandas==2.2.2
python-pptx==1.0.2
python-dotenv==1.0.1
